// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
      ai_features: {
        Row: {
          created_at: string | null
          description: string
          icon: string
          id: string
          name: string
        }
        Insert: {
          created_at?: string | null
          description: string
          icon: string
          id?: string
          name: string
        }
        Update: {
          created_at?: string | null
          description?: string
          icon?: string
          id?: string
          name?: string
        }
        Relationships: []
      }
// ...
      reservations: {
        Row: {
          client_email: string
          client_name: string
          client_phone: string
          created_at: string | null
          date: string
          id: string
          service: string
          status: string
          time: string
          updated_at: string | null
        }
// ...
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}